feat: new sidebar
