+++ package.json
@@ -20,6 +20,7 @@
     "@hookform/resolvers": "^3.10.0",
     "@prisma/client": "5.1.1",
     "@radix-ui/react-accordion": "^1.2.3",
+    "@radix-ui/react-avatar": "^1.1.3",
     "@radix-ui/react-checkbox": "^1.1.4",
     "@radix-ui/react-collapsible": "^1.1.3",
     "@radix-ui/react-dialog": "^1.1.6",
